increase data types

+++ app/inc/data.h
@@ -61,8 +61,8 @@ typedef struct CONTROL_T {
 
 typedef struct BP_LIST_T {
     char **names, **uids;
-    uint8_t hover, selected;
-    uint8_t page_min, page_max, menu_max;
+    uint16_t hover, selected;
+    uint16_t page_min, page_max, menu_max;
 } bp_list_t;
 
 typedef struct BANK_CONFIG_T {
